Improve: :recycle: Improved the code by using WP in-built get_template_parts()
This code is almost the same as our previous commit of improving reusable function code. The differences is that if we have a standard WP HTML and CSS, then we use the WP get_template_part() function. If we have custom arguments on the function we created on the previous commit, then we use that solution instead.