Improve: :sparkles: Improved better search query
Before on WYSIWYG editor. You wrote keyword math on the biology program, and when the user is going to use the search function. The problem is will get for example math on the program column which is not related to math on the biology.

To fix this, it is created by using the Advanced Custom Fields by telling the WP we can use either the main WYSIWYG editor, or using the Advanced Custom Fields where we can use that editor to filter away for example math and  biology on the keyword. This will make sure our search field will display correct result only.

The second part is on single-program.php will only display custom field based on the field name of the Advanced Custom Field