Add: :sparkles: Added a new feature for user who has disabled JavaScript on their browser, but can still use search button
If the user has disabled their JavaScript on their browser. The user can still click on the search button and will jump through the search page. On the search page, the user can search whatever in there to get the result. 

The search is working, thanks for the form. We want this feature to be a fallback to  guarantee the visitor in case if they don't have the JavaScript for example enabled on their website