Add: :sparkles: Notes is added for readonly attribute on the html tag
This is because we want the HTML to be readonly. If not, the user can click on the input and textarea element to write something. It will not even save as well. Gives a bad UX. Next stop is to make sure clicking on the edit button will let the user edit their post