Add/feature: :sparkles: Adding new notes is possible
User can add new notes title and body. It will not display on the front-end, but it will be saved on the back-end. 

In order to display it on the front end, user have to refresh the site. 

Next step is to make sure it will update the new post real-time